ABI BREAK: edgepaint: use a 'size_t' for 'lab_gamut_data_size'
Allows removal of some cast noise. This is conservatively marked as an ABI break
because liblab_gamut is shipped, though lab_gamut.h is not so it is very
unlikely this library has any external users. This change likely has no impact
for all users.

Gitlab: related to #2489

+++ lib/edgepaint/lab_gamut.h
@@ -10,6 +10,8 @@
 
 #pragma once
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -37,7 +39,7 @@ extern "C" {
  *   https://en.wikipedia.org/wiki/CIELAB_color_space
  */
 LAB_GAMUT_API extern const signed char lab_gamut_data[];
-LAB_GAMUT_API extern int lab_gamut_data_size;
+LAB_GAMUT_API extern size_t lab_gamut_data_size;
 
 #undef LAB_GAMUT_API
